+++ sdk/postgresql/Azure.ResourceManager.PostgreSql/src/Custom/PostgreSqlFlexibleServer/PostgreSqlFlexibleServerMicrosoftEntraAdministratorData.cs
@@ -1,0 +1,15 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+#nullable disable
+
+namespace Azure.ResourceManager.PostgreSql.FlexibleServers
+{
+    public partial class PostgreSqlFlexibleServerMicrosoftEntraAdministratorData
+    {
+        /// <summary> Initializes a new instance of <see cref="PostgreSqlFlexibleServerMicrosoftEntraAdministratorData"/>. </summary>
+        public PostgreSqlFlexibleServerMicrosoftEntraAdministratorData()
+        {
+        }
+    }
+}